requirements.txt
Django==4.2
joblib~=1.2.0
numpy~=1.24.2
scikit-learn~=1.2.2
pandas==2.0.0
python-dotenv==1.0.0
django-compressor
pyodbc --no-binary=pyodbc